# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/likes": {
            "post": {
                "description": "Добавляет лайк к фото от имени пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Likes"
                ],
                "summary": "Добавить лайк",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID фото",
                        "name": "photoID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "userID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: Like added successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректные параметры",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет лайк с фото, поставленный пользователем",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Likes"
                ],
                "summary": "Удалить лайк",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID фото",
                        "name": "photoID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID пользователя",
                        "name": "userID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: Like removed successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректные параметры",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/likes/count": {
            "get": {
                "description": "Возвращает количество лайков у фото",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Likes"
                ],
                "summary": "Получить количество лайков",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID фото",
                        "name": "photoID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "likes_count: Количество лайков",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Некорректные параметры",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/likes/users": {
            "get": {
                "description": "Возвращает список пользователей, поставивших лайк на фото",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Likes"
                ],
                "summary": "Получить список лайков",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID фото",
                        "name": "photoID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "users: [список пользователей]",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Некорректные параметры",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/messages": {
            "post": {
                "description": "Отправляет новое сообщение в указанную беседу",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Отправить сообщение",
                "parameters": [
                    {
                        "description": "Данные сообщения",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message_id: ID созданного сообщения",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/messages/{conversationID}": {
            "get": {
                "description": "Возвращает список сообщений по conversation_id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Получить сообщения",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID беседы",
                        "name": "conversationID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Message"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный conversation_id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/messages/{messageID}": {
            "delete": {
                "description": "Удаляет сообщение по его ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Удалить сообщение",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID сообщения",
                        "name": "messageID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: Message deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Некорректный ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "Содержимое сообщения",
                    "type": "string",
                    "example": "Привет! Как дела?"
                },
                "conversation_id": {
                    "description": "ID беседы, к которой относится сообщение",
                    "type": "integer",
                    "example": 101
                },
                "created_at": {
                    "description": "Дата и время отправки сообщения",
                    "type": "string",
                    "example": "2024-02-01T15:45:00Z"
                },
                "id": {
                    "description": "ID сообщения",
                    "type": "integer",
                    "example": 1
                },
                "sender_id": {
                    "description": "ID отправителя сообщения",
                    "type": "integer",
                    "example": 42
                }
            }
        },
...
    }
}